+++ scripts/experiment_first_step_floor.py
@@ -21,7 +21,9 @@ from experiment_horizon_generalization import (
     write,
 )
 
-from glassbox.experimental.default_model import _RECIPE, LearnedDynamics, fit
+from glassbox.experimental.default_model import _HISTORY_RECIPE as _RECIPE
+from glassbox.experimental.default_model import LearnedDynamics
+from glassbox.experimental.default_model import _fit_history as fit
 from glassbox.experimental.sequence_model import fit_sequence_model