+++ scripts/experiment_history_confounding.py
@@ -19,7 +19,9 @@ import numpy as np
 from experiment_model_qualification import collection, encoding, recordings
 from experiment_sequence_diagnostics import fresh_recordings
 
-from glassbox.experimental.default_model import _RECIPE, LearnedDynamics, fit
+from glassbox.experimental.default_model import _HISTORY_RECIPE as _RECIPE
+from glassbox.experimental.default_model import LearnedDynamics
+from glassbox.experimental.default_model import _fit_history as fit
 from glassbox.experimental.sequence_collection import (
     SequenceCollection,
     SequenceSegment,